Move AI reformatting plugin to submodule (#17)
* Move the AI plugin into a submodule

- Creates localgov_publications_importer_ai submodule
- Move the AI transform plugin into this module
- Remove AI as a dependency on the main module

* Move the drupal/ai dependency to composer suggest

* Update readme (AI submodule references)

Add information about the new AI submodule and indicate that this module and an AI
provider needs to be installed in order to use AI to reformat the PDF.

* Amend composer suggests text

Add that drupal/ai is required by the localgov_publications_importer_ai submodule.

* Remove debug lines

* Move position of use statement to correct order

* Update readme with other AI provider references

In official BBC style voice: "Other AI LLM providers are also available."
The link goes to the AI module page where it lists other provider modules.
Also documents where to get an Open AI api key from.

* Remove Open AI php library dependency

This will be brought in if a user requires drupal/ai_provider_open_ai